Update react and react types (#1884)
* update react types version to v18

The first version of react-query that is compatible with the new major
version of react types is 3.35.0. I updated all the dev dependencies
accordingly to fix the tests. Technically the peer dependencies should
not be affacted but just to be safe I updated them as well so that the
dev dependencies aren't higher than the peer dependencies.

* update react and react-dom

* fix react types version for testing library

@testing-library/react has a transitive dependency on @types/react@*.
The resolved version in yarn.lock was still v17 which leads to build
errors. The yarn.lock file was manually fixed.